:sparkles: Refactored ScriptField to extract a ScriptTrait
